Rename file
To match the name of the default export
